Generate constant regex for match with a constant pattern.
For constraints like match(".*", X), where the pattern is a
string constant, we can avoid using the regex cache. Avoiding the
cache is going to perform better in all cases.

I introduced the new node RegexConstant which only appears immediately
below a MATCH or NOT_MATCH node.

The behavior with regards to invalid regexes has not changed and bad
regexes will not lead to program errors.