Fix source locations with `--no-preprocessor` and add a virtual file system layer to the parser (#2377)
* Add a virtual file system layer to the parser

The default virtual file system used by the parser is the real
file system. And it's possible to initialize the parser driver
with a different virtual file system.

The Overlay file system is a stack of virtual file systems.

* Fix source locations with `--no-preprocessor`.

Source location were off by one line.

The source location on the first line had an uninitialized column
number.

Error messages would point to the wrong column when the location is
after consecutive non-leading whitespaces.